requirements.txt
accelerate==1.10.0
acres==0.5.0
addict==2.4.0
aiofiles==23.2.1
airportsdata==20250811
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
async-timeout==4.0.3
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.13.4
boto3==1.40.8
botocore==1.40.8
bottle==0.13.4
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
common==0.1.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
data==0.4
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
doclayout_yolo==0.0.2b1
dual==0.1.0
dynamo3==0.4.10
einops==0.8.1
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fast-langdetect==0.3.2
fastapi==0.116.1
fasttext-predict==0.9.2.4
ffmpy==0.6.1
filelock==3.18.0
fire==0.7.0
fitz==0.0.1.dev2
flywheel==0.5.4
fonttools==4.59.0
frontend==0.0.3
fsspec==2025.7.0
funcsigs==1.0.2
genson==1.3.0
gradio==5.23.3
gradio_client==1.8.0
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.5.2
iso3166==2.1.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langsmith==0.4.13
lmdeploy==0.8.0
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mmengine-lite==0.10.7
modelscope==1.28.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==1.26.4
nvidia-ml-py==12.575.51
openai==1.88.0
opencv-contrib-python==4.12.0.88
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.1
outlines==1.2.3
outlines_core==0.2.11
packaging==25.0
paddlepaddle==3.1.0
paddlex==3.1.3
pandas==2.3.1
partial-json-parser==0.2.1.1.post6
pathlib==1.0.1
pdfminer.six==20231228
peewee==3.18.2
peft==0.14.0
pillow==11.3.0
platformdirs==4.3.8
ply==3.11
prettytable==3.16.0
protobuf==3.20.2
prov==2.1.1
prox==0.0.17
psutil==7.0.0
puremagic==1.30
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.24.14
pynvml==12.0.0
pyparsing==3.2.3
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-geoip-python3==1.3
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
qwen-vl-utils==0.0.10
ray==2.48.0
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
robust-downloader==0.0.2
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.8
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.20.1
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
stringzilla==3.12.6
sympy==1.13.1
tenacity==9.1.2
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tight==0.1.0
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.3
torchvision==0.20.1
tqdm==4.67.1
traits==7.0.2
transformers==4.52.4
triton-windows==3.4.0.post20
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
xlsxwriter==3.2.5
yapf==0.43.0
zstandard==0.23.0
